moved strand class to another file and improved readability
